% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_meritada.R
\name{air_meritada}
\alias{air_meritada}
\title{Càlcul de la despesa meritada per AIR (Atenció a la Insuficiència Renal)}
\usage{
air_meritada(QC, tarifa, QF)
}
\arguments{
\item{QC}{Quantitat contractada.}

\item{tarifa}{Tarifa unitària.}

\item{QF}{Quantitat facturada/realitzada.}
}
\value{
Un numèric amb la despesa meritada agregada.
}
\description{
Càlcul de la despesa meritada per AIR (Atenció a la Insuficiència Renal)
}
\examples{
air_meritada(QC = c(200, 150), tarifa = c(90, 110), QF = c(210, 140))
}
