% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bombes_insulina_meritada.R
\name{bombes_insulina_meritada}
\alias{bombes_insulina_meritada}
\title{Càlcul de la despesa meritada per bombes d'insulina}
\usage{
bombes_insulina_meritada(QC, tarifa, QF)
}
\arguments{
\item{QC}{Quantitat contractada.}

\item{tarifa}{Tarifa unitària.}

\item{QF}{Quantitat facturada/realitzada.}
}
\value{
Un numèric amb la despesa meritada agregada.
}
\description{
Càlcul de la despesa meritada per bombes d'insulina
}
\examples{
bombes_insulina_meritada(QC = c(40, 30), tarifa = c(3500, 4000), QF = c(38, 28))
}
