% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma_complexa_meritada.R
\name{cma_complexa_meritada}
\alias{cma_complexa_meritada}
\title{Càlcul de la despesa meritada per CMA complexa}
\usage{
cma_complexa_meritada(QC, tarifa, QF)
}
\arguments{
\item{QC}{Quantitat contractada.}

\item{tarifa}{Tarifa unitària.}

\item{QF}{Quantitat facturada/realitzada.}
}
\value{
Un numèric amb la despesa meritada agregada.
}
\description{
Càlcul de la despesa meritada per CMA complexa
}
\examples{
cma_complexa_meritada(QC = c(50, 30), tarifa = c(800, 600), QF = c(55, 28))
}
