% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospi.R
\name{hospi}
\alias{hospi}
\title{Càlcul de la despesa meritada per hospitalització (AH)}
\usage{
hospi(QC, tarifa, QF, reingressos)
}
\arguments{
\item{QC}{Quantitat contractada.}

\item{tarifa}{Tarifa unitària.}

\item{QF}{Quantitat facturada/realitzada.}

\item{reingressos}{Nombre de reingressos.}
}
\value{
Vector numèric amb la despesa meritada.
}
\description{
Càlcul de la despesa meritada per hospitalització (AH)
}
\examples{
hospi(QC = 100, tarifa = 500, QF = 110, reingressos = 2)
}
