% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital_dia_cir_menor_meritada.R
\name{hospital_dia_cir_menor_meritada}
\alias{hospital_dia_cir_menor_meritada}
\title{Càlcul de la despesa meritada per hospital de dia / cirurgia menor}
\usage{
hospital_dia_cir_menor_meritada(QC, tarifa, QF)
}
\arguments{
\item{QC}{Quantitat contractada.}

\item{tarifa}{Tarifa unitària.}

\item{QF}{Quantitat facturada/realitzada.}
}
\value{
Vector numèric amb la despesa meritada.
}
\description{
Càlcul de la despesa meritada per hospital de dia / cirurgia menor
}
\examples{
hospital_dia_cir_menor_meritada(QC = 100, tarifa = 300, QF = 108)
}
