% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laf_meritada.R
\name{laf_meritada}
\alias{laf_meritada}
\title{Càlcul de la despesa meritada per LAF}
\usage{
laf_meritada(tarifa, QF)
}
\arguments{
\item{tarifa}{Tarifa unitària.}

\item{QF}{Quantitat facturada/realitzada.}
}
\value{
Vector numèric amb la despesa meritada.
}
\description{
Càlcul de la despesa meritada per LAF
}
\examples{
laf_meritada(tarifa = 45, QF = 200)
}
