% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num0}
\alias{num0}
\title{Substitueix NA per 0}
\usage{
num0(x)
}
\arguments{
\item{x}{Valor numèric (longitud 1).}
}
\value{
`x`, o `0` si `x` és `NA`.
}
\description{
Substitueix NA per 0
}
\keyword{internal}
