% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncologia_mastectomia_meritada.R
\name{oncologia_mastectomia_meritada}
\alias{oncologia_mastectomia_meritada}
\title{Càlcul de la despesa meritada per oncologia de precisió (mastectomia)}
\usage{
oncologia_mastectomia_meritada(QC, tarifa, QF)
}
\arguments{
\item{QC}{Quantitat contractada.}

\item{tarifa}{Tarifa unitària.}

\item{QF}{Quantitat facturada/realitzada.}
}
\value{
Un numèric amb la despesa meritada agregada.
}
\description{
Càlcul de la despesa meritada per oncologia de precisió (mastectomia)
}
\examples{
oncologia_mastectomia_meritada(QC = c(30, 20), tarifa = c(2500, 1800), QF = c(33, 19))
}
