% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protesis_meritada.R
\name{protesis_meritada}
\alias{protesis_meritada}
\title{Càlcul de la despesa meritada per pròtesis}
\usage{
protesis_meritada(CS, PM, tarifa, QF)
}
\arguments{
\item{CS}{Nombre de casos contractats (longitud 1).}

\item{PM}{Preu mig contractat (longitud 1).}

\item{tarifa}{Tarifa unitària per cas.}

\item{QF}{Quantitat facturada/realitzada per cas.}
}
\value{
Un numèric amb la despesa meritada.
}
\description{
Càlcul de la despesa meritada per pròtesis
}
\examples{
protesis_meritada(CS = 100, PM = 3000, tarifa = c(2800, 3200), QF = c(60, 45))
}
