% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rehabilitacio_meritada.R
\name{rehabilitacio_meritada}
\alias{rehabilitacio_meritada}
\title{Càlcul de la despesa meritada per rehabilitació}
\usage{
rehabilitacio_meritada(QC, tarifa, QF)
}
\arguments{
\item{QC}{Quantitat contractada.}

\item{tarifa}{Tarifa unitària.}

\item{QF}{Quantitat facturada/realitzada.}
}
\value{
Un numèric amb la despesa meritada agregada.
}
\description{
Càlcul de la despesa meritada per rehabilitació
}
\examples{
rehabilitacio_meritada(QC = c(500, 300), tarifa = c(20, 25), QF = c(520, 310))
}
