% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reingressos_meritada.R
\name{reingressos_meritada}
\alias{reingressos_meritada}
\title{Càlcul de la despesa meritada per reingressos}
\usage{
reingressos_meritada(
  QF,
  reingressos,
  tarifa_alta,
  tarifa_reingres,
  pct_reingressos_pactat,
  pct_reingressos_limit
)
}
\arguments{
\item{QF}{Quantitat facturada/realitzada.}

\item{reingressos}{Nombre de reingressos.}

\item{tarifa_alta}{Tarifa d'alta.}

\item{tarifa_reingres}{Tarifa de reingrés.}

\item{pct_reingressos_pactat}{Percentatge de reingressos pactat.}

\item{pct_reingressos_limit}{Percentatge de reingressos límit.}
}
\value{
Vector numèric amb la despesa meritada.
}
\description{
Càlcul de la despesa meritada per reingressos
}
\examples{
reingressos_meritada(
  QF = 100, reingressos = 8, tarifa_alta = 3000, tarifa_reingres = 3000,
  pct_reingressos_pactat = 0.05, pct_reingressos_limit = 0.10
)
}
