% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduccio_humana_assistida_meritada.R
\name{reproduccio_humana_assistida_meritada}
\alias{reproduccio_humana_assistida_meritada}
\title{Càlcul de la despesa meritada per reproducció humana assistida}
\usage{
reproduccio_humana_assistida_meritada(QC, tarifa, QF)
}
\arguments{
\item{QC}{Quantitat contractada.}

\item{tarifa}{Tarifa unitària.}

\item{QF}{Quantitat facturada/realitzada.}
}
\value{
Un numèric amb la despesa meritada agregada.
}
\description{
Càlcul de la despesa meritada per reproducció humana assistida
}
\examples{
reproduccio_humana_assistida_meritada(QC = c(20, 10), tarifa = c(1200, 900), QF = c(22, 9))
}
