% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_excel}
\alias{round_excel}
\title{Arrodoniment estil Excel}
\usage{
round_excel(x, digits = 2)
}
\arguments{
\item{x}{Vector numèric a arrodonir.}

\item{digits}{Nombre de decimals (per defecte 2).}
}
\value{
Vector numèric arrodonit.
}
\description{
Arrodoneix un valor numèric de la mateixa manera que ho fa Excel
(arrodoniment "half away from zero"), eliminant primer el soroll de
coma flotant típic de R (p. ex. 984.7349999999999 -> 984.735).
}
\keyword{internal}
