% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttpe_meritada.R
\name{ttpe_meritada}
\alias{ttpe_meritada}
\title{Càlcul de la despesa meritada per TTPE}
\usage{
ttpe_meritada(QC, tarifa, QF, codi_activitat)
}
\arguments{
\item{QC}{Quantitat contractada.}

\item{tarifa}{Tarifa unitària.}

\item{QF}{Quantitat facturada/realitzada.}

\item{codi_activitat}{Codi d'activitat (character), ha d'existir a la
taula interna de percentatges.}
}
\value{
Vector numèric amb la despesa meritada.
}
\description{
Aplica la taula de percentatges per `codi_activitat` (extreta de
"Tarifes Aguts") a cada cas.
}
\examples{
ttpe_meritada(QC = 100, tarifa = 400, QF = 110, codi_activitat = "143205")
}
