% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urgencies_atesa_meritada.R
\name{urgencies_atesa_meritada}
\alias{urgencies_atesa_meritada}
\title{Càlcul de la despesa meritada per urgències (atesa)}
\usage{
urgencies_atesa_meritada(QC, tarifa, QF)
}
\arguments{
\item{QC}{Quantitat contractada.}

\item{tarifa}{Tarifa unitària.}

\item{QF}{Quantitat facturada/realitzada.}
}
\value{
Vector numèric amb la despesa meritada.
}
\description{
Càlcul de la despesa meritada per urgències (atesa)
}
\examples{
urgencies_atesa_meritada(QC = 100, tarifa = 50, QF = 105)
}
