% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urgencies_hivern_meritada.R
\name{urgencies_hivern_meritada}
\alias{urgencies_hivern_meritada}
\title{Càlcul de la despesa meritada per urgències d'hivern}
\usage{
urgencies_hivern_meritada(QC, tarifa, QF)
}
\arguments{
\item{QC}{Quantitat contractada.}

\item{tarifa}{Tarifa unitària.}

\item{QF}{Quantitat facturada/realitzada.}
}
\value{
Un numèric amb la despesa meritada agregada.
}
\description{
Càlcul de la despesa meritada per urgències d'hivern
}
\examples{
urgencies_hivern_meritada(QC = c(300, 200), tarifa = c(60, 55), QF = c(320, 190))
}
