% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urgencies_triatge_meritada.R
\name{urgencies_triatge_meritada}
\alias{urgencies_triatge_meritada}
\title{Càlcul de la despesa meritada per urgències (triatge)}
\usage{
urgencies_triatge_meritada(QC, tarifa, QF)
}
\arguments{
\item{QC}{Quantitat contractada.}

\item{tarifa}{Tarifa unitària.}

\item{QF}{Quantitat facturada/realitzada.}
}
\value{
Vector numèric amb la despesa meritada.
}
\description{
Càlcul de la despesa meritada per urgències (triatge)
}
\examples{
urgencies_triatge_meritada(QC = 100, tarifa = 50, QF = 130)
}
